feat：线索流程完成可以查询线索，公开线索，实现了调用接口可以查询自己的隐藏线索和公开线索，用中间表作为媒介

+++ script-murder/src/main/java/com/wn/entity/dm/RoomCluePO.java
@@ -37,7 +37,7 @@ public class RoomCluePO {
     private LocalDateTime obtainTime;  ///领取时间，记录玩家搜到这条线索的时间
 
     @Column(name = "is_public")
-    private Byte isPublic = 0;  ///是否公开，0 表示不公开，1 表示公开 ，默认0
+    private int isPublic = 0;  ///是否公开，0 表示不公开，1 表示公开 ，默认0
 
     @PrePersist
     public void prePersist() {
